requirements.txt
alembic==1.8.1
amqp==5.1.1
anyio==3.6.1
async-timeout==4.0.2
bcrypt==4.0.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.0
fastapi-pagination==0.10.0
h11==0.13.0
httptools==0.5.0
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.9.0
install==1.3.5
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.3
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
prompt-toolkit==3.0.31
psycopg2==2.9.3
psycopg2-binary==2.9.3
pydantic==1.10.2
PyJWT==2.5.0
pyparsing==3.0.9
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.3.0
uvicorn==0.18.3
uvloop==0.17.0
vine==5.0.0
watchfiles==0.17.0
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.1
zipp==3.8.1
